refactor(metrics): separate http server and metrics into two modules
